refactor(rust): Split `parquet_source.rs` in new-streaming
